backend/requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
neo4j>=5.14.0
pandas>=2.2.0
numpy>=1.26.0
networkx>=3.2.0
python-dotenv>=1.0.0
gunicorn>=21.2.0

